financial_audit_env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-audit-env"
version = "1.0.0"
description = "OpenEnv-compatible RL environment for financial auditing tasks"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Harshit"},
]
keywords = ["openenv", "reinforcement-learning", "financial-audit", "ai-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
openenv = ["openenv-core>=0.2.0"]
dev = ["pytest>=7.0", "httpx>=0.25.0"]

[project.urls]
Homepage = "https://github.com/Harshit/financial-audit-env"
Repository = "https://github.com/Harshit/financial-audit-env"

[tool.setuptools.packages.find]
include = ["financial_audit_env*"]
